write tcpdump file for each scenario
